Re-calibrate classification probability predictions

Calibration is the process of adjusting a model's outputted probabilities to match the observed frequencies of events. This technique aims to ensure that when a model predicts a certain probability for an outcome, that probability accurately reflects the true likelihood of that outcome occurring.

Usage

adjust_probability_calibration(x, method = NULL, ...)

Arguments

x

A tailor().

method

Character. One of "logistic", "multinomial", "beta", "isotonic", or "isotonic_boot", corresponding to the function from the probably package probably::cal_estimate_logistic(), probably::cal_estimate_multinomial(), etc., respectively.

...

Optional arguments to pass to the corresponding function in the probably package. These arguments must be named.

Data Usage

This adjustment requires estimation and, as such, different subsets of data should be used to train it and evaluate its predictions.

Examples

library(modeldata)

# split example data
set.seed(1)
in_rows <- sample(c(TRUE, FALSE), nrow(two_class_example), replace = TRUE)
d_calibration <- two_class_example[in_rows, ]
d_test <- two_class_example[!in_rows, ]

head(d_calibration)
#>    truth      Class1       Class2 predicted
#> 1 Class2 0.003589243 0.9964107574    Class2
#> 3 Class2 0.110893522 0.8891064779    Class2
#> 4 Class1 0.735161703 0.2648382969    Class1
#> 6 Class1 0.999275071 0.0007249286    Class1
#> 7 Class1 0.999201149 0.0007988510    Class1
#> 8 Class1 0.812351997 0.1876480026    Class1

# specify calibration
tlr <-
  tailor() |>
  adjust_probability_calibration(method = "logistic")

# train tailor on a subset of data.
tlr_fit <- fit(
  tlr,
  d_calibration,
  outcome = c(truth),
  estimate = c(predicted),
  probabilities = c(Class1, Class2)
)

# apply to predictions on another subset of data
head(d_test)
#>     truth       Class1      Class2 predicted
#> 2  Class1 0.6786210540 0.321378946    Class1
#> 5  Class2 0.0162399603 0.983760040    Class2
#> 9  Class2 0.4570372595 0.542962741    Class2
#> 10 Class2 0.0976377342 0.902362266    Class2
#> 16 Class1 0.9919162773 0.008083723    Class1
#> 17 Class2 0.0004250834 0.999574917    Class2

predict(tlr_fit, d_test)
#> # A tibble: 244 × 4
#>    truth     Class1    Class2 predicted
#>    <fct>  <dbl[1d]> <dbl[1d]> <fct>    
#>  1 Class1    0.459     0.541  Class2   
#>  2 Class2    0.0396    0.960  Class2   
#>  3 Class2    0.333     0.667  Class2   
#>  4 Class2    0.0954    0.905  Class2   
#>  5 Class1    0.964     0.0359 Class1   
#>  6 Class2    0.0328    0.967  Class2   
#>  7 Class2    0.112     0.888  Class2   
#>  8 Class1    0.665     0.335  Class1   
#>  9 Class1    0.957     0.0434 Class1   
#> 10 Class2    0.0534    0.947  Class2   
#> # ℹ 234 more rows